Map Tile Scripting
Copied the Location Text for all 216 map tiles to the appropriate MapTileScriptable asset, along with the other values for which I've already implemented the "rules" (like rolling for multiple ranges, allowing the player to Move On, the time taken, XP earned, etc).

Probably did some other stuff too, but this process was so time-consuming and tedious I've forgotten.