Implementing Various Encounter Rules
Began implementing rules for progressing through encounters, including:
(A lot of these are similar to the same rules on the Location Texts, so they only needed some copy-paste or slight tweaks).

-Multiple Ranges for random selection of progress
-Adding skills to the random selection roll per the encounter text's instruction

-"If you roll within a certain range, you die. Else, go to [specific encounter]."

-Basic Win Screen UI background, with Animator.
-Basic Game Over Screen UI background, with Animator.

-Automatically Moving On to [specific tile], per the encounter text's instruction, and from there, taking the player to the relevant Location text, or allowing them to move on.
-Automatically Moving On in a random direction, per the encounter text's instruction.

Committing and pushing this now, so I can look it up at work and try to get some of this done while I'm there.